exeptions/memMapped
ok so exeptions i found online that has interupts for a bunch of
different things make sure you use the keyboard and display simulator in
tools or it wont work. You have to connected it and sometimes its buggy
so try hitting reset if nothing comes up.
memMapped is basically the same thing but i cut out all the stuff that
wasnt keyboard interupts
this has a pretty good explenation on the basic idea of excaptions and
memory mapped io
http://www.cs.nott.ac.uk/~txa/g51csa/l12-hand.pdf
this ones not bad either
https://courses.engr.illinois.edu/cs232/fa2011/section/disc5.pdf