Start putting the things we have worked on together
The game "functions" for the most part
the code needs a lot of cleaning up most on my end i have a lot of
random prints or mains that i will need to get rid of. I did not put in
the timer stuff yet. Some formating stuff will need to be done just to
make the game look better and function better for the user. Im probably
forgetting something and but its late and I want to go to bed. Happy
Thanksgiving everyone